requirements.txt
Django==5.1.1
celery==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
flower==2.0.1
pytest==8.3.3
redis==5.1.1
django-redis==5.4.0

